[Feat/DetailPage] 게시글 작성자인지 아닌지 판단하는 함수 분리
