+++ src/components/product-detail-page/pdp-right/elements/SmallUI/Title.tsx
@@ -1,13 +1,10 @@
 "use client";
 
 // import { useState } from "react";
-import { useState, useCallback } from "react";
 import styled from "styled-components";
-import axios from "axios";
-import { serverIP } from "@/../secrets.json";
-import GetToken from "@/utils/getToken";
 import ToolBoxForWriter from "./ToolBoxForWriter";
 import ToolBoxForGuest from "./ToolBoxForGuest";
+import useCheckIfWriter from "../CheckIfWriter";
 
 const NameBox = styled.div`
   display: flex;
@@ -23,24 +20,7 @@ const NameBox = styled.div`
 
 // 상세페이지 제목, 그 옆의 도구들 (찜, 공유, 신고)
 export default function Title({ title, id }: { title: string; id: number }) {
-  const token = GetToken();
-  const [isWriter, setIsWriter] = useState<boolean>();
-
-  // 글쓴이인지 아닌지 판단.
-  const getWriterStatus = useCallback(async () => {
-    try {
-      const response = await axios.get(`${serverIP}/api/post/validate`, {
-        headers: { Authorization: `Bearer ${token}` },
-        params: { id: id },
-      });
-      setIsWriter(response.data);
-      // console.log("찜여부:", response.data);
-    } catch (error) {
-      console.log("An error occurred while getting LikeStatus. ", error);
-    }
-  }, [id, token]);
-
-  getWriterStatus();
+  const isWriter = useCheckIfWriter({ id });
 
   return (
     <NameBox>
